refactor(parser): extract route parsing into dedicated sub-parsers
