+++ fox_test.go
@@ -2195,1272 +2195,6 @@ func TestUpdateRoute(t *testing.T) {
 	}
 }
 
-func TestParseRoute(t *testing.T) {
-	f := MustRouter(AllowRegexpParam(true))
-
-	staticToken := func(v string, hsplit bool) token {
-		return token{
-			value:  v,
-			typ:    nodeStatic,
-			hsplit: hsplit,
-		}
-	}
-
-	paramToken := func(v, reg string) token {
-		tk := token{
-			value: v,
-			typ:   nodeParam,
-		}
-		if reg != "" {
-			tk.regexp = regexp.MustCompile("^" + reg + "$")
-		}
-		return tk
-	}
-
-	wildcardToken := func(v, reg string) token {
-		tk := token{
-			value: v,
-			typ:   nodeWildcard,
-		}
-		if reg != "" {
-			tk.regexp = regexp.MustCompile("^" + reg + "$")
-		}
-		return tk
-	}
-
-	cases := []struct {
-		wantErr           error
-		name              string
-		path              string
-		wantN             int
-		wantTokens        []token
-		wantStartCatchAll int
-	}{
-		{
-			name:       "valid static route",
-			path:       "/foo/bar",
-			wantTokens: slices.Collect(iterutil.SeqOf(staticToken("/foo/bar", false))),
-		},
-		{
-			name:  "top level domain param",
-			path:  "{tld}/foo/bar",
-			wantN: 1,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				paramToken("tld", ""),
-				staticToken("/foo/bar", false),
-			)),
-		},
-		{
-			name:  "top level domain wildcard",
-			path:  "+{tld}/foo/bar",
-			wantN: 1,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				wildcardToken("tld", ""),
-				staticToken("/foo/bar", false),
-			)),
-		},
-		{
-			name:  "valid catch all route",
-			path:  "/foo/bar/+{arg}",
-			wantN: 1,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/bar/", false),
-				wildcardToken("arg", ""),
-			)),
-			wantStartCatchAll: 9,
-		},
-		{
-			name:  "valid param route",
-			path:  "/foo/bar/{baz}",
-			wantN: 1,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/bar/", false),
-				paramToken("baz", ""),
-			)),
-		},
-		{
-			name:  "valid multi params route",
-			path:  "/foo/{bar}/{baz}",
-			wantN: 2,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/", false),
-				paramToken("bar", ""),
-				staticToken("/", false),
-				paramToken("baz", ""),
-			)),
-		},
-		{
-			name:  "valid same params route",
-			path:  "/foo/{bar}/{bar}",
-			wantN: 2,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/", false),
-				paramToken("bar", ""),
-				staticToken("/", false),
-				paramToken("bar", ""),
-			)),
-		},
-		{
-			name:  "valid multi params and catch all route",
-			path:  "/foo/{bar}/{baz}/+{arg}",
-			wantN: 3,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/", false),
-				paramToken("bar", ""),
-				staticToken("/", false),
-				paramToken("baz", ""),
-				staticToken("/", false),
-				wildcardToken("arg", ""),
-			)),
-			wantStartCatchAll: 17,
-		},
-		{
-			name:  "valid inflight param",
-			path:  "/foo/xyz:{bar}",
-			wantN: 1,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/xyz:", false),
-				paramToken("bar", ""),
-			)),
-		},
-		{
-			name:  "valid inflight catchall",
-			path:  "/foo/xyz:+{bar}",
-			wantN: 1,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/xyz:", false),
-				wildcardToken("bar", ""),
-			)),
-			wantStartCatchAll: 9,
-		},
-		{
-			name:  "valid multi inflight param and catch all",
-			path:  "/foo/xyz:{bar}/abc:{bar}/+{arg}",
-			wantN: 3,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/xyz:", false),
-				paramToken("bar", ""),
-				staticToken("/abc:", false),
-				paramToken("bar", ""),
-				staticToken("/", false),
-				wildcardToken("arg", ""),
-			)),
-			wantStartCatchAll: 25,
-		},
-		{
-			name:  "catch all with arg in the middle of the route",
-			path:  "/foo/bar/+{bar}/baz",
-			wantN: 1,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/bar/", false),
-				wildcardToken("bar", ""),
-				staticToken("/baz", false),
-			)),
-		},
-		{
-			name:  "multiple catch all suffix and inflight with arg in the middle of the route",
-			path:  "/foo/bar/+{bar}/x+{args}/y/+{z}/{b}",
-			wantN: 4,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/bar/", false),
-				wildcardToken("bar", ""),
-				staticToken("/x", false),
-				wildcardToken("args", ""),
-				staticToken("/y/", false),
-				wildcardToken("z", ""),
-				staticToken("/", false),
-				paramToken("b", ""),
-			)),
-		},
-		{
-			name:  "inflight catch all with arg in the middle of the route",
-			path:  "/foo/bar/damn+{bar}/baz",
-			wantN: 1,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/bar/damn", false),
-				wildcardToken("bar", ""),
-				staticToken("/baz", false),
-			)),
-		},
-		{
-			name:  "catch all with arg in the middle of the route and param after",
-			path:  "/foo/bar/+{bar}/{baz}",
-			wantN: 2,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/bar/", false),
-				wildcardToken("bar", ""),
-				staticToken("/", false),
-				paramToken("baz", ""),
-			)),
-		},
-		{
-			name:  "simple domain and path",
-			path:  "foo/bar",
-			wantN: 0,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("foo", true),
-				staticToken("/bar", false),
-			)),
-		},
-		{
-			name:  "simple domain with trailing slash",
-			path:  "foo/",
-			wantN: 0,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("foo", true),
-				staticToken("/", false),
-			)),
-		},
-		{
-			name:  "period in param path allowed",
-			path:  "foo/{.bar}",
-			wantN: 1,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("foo", true),
-				staticToken("/", false),
-				paramToken(".bar", ""),
-			)),
-		},
-		{
-			name:    "missing a least one slash",
-			path:    "foo.com",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "empty parameter",
-			path:    "/foo/bar{}",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "missing arguments name after catch all",
-			path:    "/foo/bar/*",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "missing arguments name after param",
-			path:    "/foo/bar/{",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "catch all in the middle of the route",
-			path:    "/foo/bar/*/baz",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "empty infix catch all",
-			path:    "/foo/bar/+{}/baz",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "empty ending catch all",
-			path:    "/foo/bar/baz/+{}",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "unexpected character in param",
-			path:    "/foo/{{bar}",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "unexpected character in param",
-			path:    "/foo/{*bar}",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "unexpected character in catch-all",
-			path:    "/foo/+{/bar}",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "catch all not supported in hostname",
-			path:    "a.b.c*/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "illegal character in params hostname",
-			path:    "a.b.c{/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "illegal character in hostname label",
-			path:    "a.b.c}/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "unexpected character in param hostname",
-			path:    "a.{.bar}.c/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "unexpected character in wildcard hostname",
-			path:    "a.+{.bar}.c/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "unexpected character in param hostname",
-			path:    "a.{/bar}.c/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "unexpected character in wildcard hostname",
-			path:    "a.+{/bar}.c/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "in flight catch-all after param in one route segment",
-			path:    "/foo/{bar}+{baz}",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "multiple param in one route segment",
-			path:    "/foo/{bar}{baz}",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "in flight param after catch all",
-			path:    "/foo/+{args}{param}",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "consecutive catch all with no slash",
-			path:    "/foo/+{args}+{param}",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "consecutive catch all",
-			path:    "/foo/+{args}/+{param}",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "consecutive catch all with inflight",
-			path:    "/foo/ab+{args}/+{param}",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "unexpected char after inflight catch all",
-			path:    "/foo/ab+{args}a",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "unexpected char after catch all",
-			path:    "/foo/+{args}a",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:  "prefix catch-all in hostname",
-			path:  "+{any}.com/foo",
-			wantN: 1,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				wildcardToken("any", ""),
-				staticToken(".com", true),
-				staticToken("/foo", false),
-			)),
-		},
-		{
-			name:  "infix catch-all in hostname",
-			path:  "a.+{any}.com/foo",
-			wantN: 1,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("a.", true),
-				wildcardToken("any", ""),
-				staticToken(".com", true),
-				staticToken("/foo", false),
-			)),
-		},
-		{
-			name:  "illegal catch-all in hostname",
-			path:  "a.b.+{any}/foo",
-			wantN: 1,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("a.b.", true),
-				wildcardToken("any", ""),
-				staticToken("/foo", false),
-			)),
-		},
-		{
-			name:  "static hostname with catch-all path",
-			path:  "a.b.com/+{any}",
-			wantN: 1,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("a.b.com", true),
-				staticToken("/", false),
-				wildcardToken("any", ""),
-			)),
-			wantStartCatchAll: 8,
-		},
-		{
-			name:    "illegal control character in path",
-			path:    "example.com/foo\x00",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "illegal leading hyphen in hostname",
-			path:    "-a.com/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "illegal leading dot in hostname",
-			path:    ".a.com/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "illegal trailing hyphen in hostname",
-			path:    "a.com-/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "illegal trailing dot in hostname",
-			path:    "a.com./",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "illegal trailing dot in hostname after param",
-			path:    "{tld}./foo/bar",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "illegal single dot in hostname",
-			path:    "./",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "illegal hyphen before dot",
-			path:    "a-.com/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "illegal hyphen after dot",
-			path:    "a.-com/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "illegal double dot",
-			path:    "a..com/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "illegal double dot with param state",
-			path:    "{b}..com/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "illegal double dot with inflight param state",
-			path:    "a{b}..com/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "param not finishing with delimiter in hostname",
-			path:    "{a}b{b}.com/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "consecutive parameter in hostname",
-			path:    "{a}{b}.com/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "leading hostname label exceed 63 characters",
-			path:    "uj01dowf1x5lk6lysurbr0lgbdd1wfyw8sm8q17mnt0i9igk774vcwr5rly5dguu.b.com/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "middle hostname label exceed 63 characters",
-			path:    "a.uj01dowf1x5lk6lysurbr0lgbdd1wfyw8sm8q17mnt0i9igk774vcwr5rly5dguu.com/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "trailing hostname label exceed 63 characters",
-			path:    "a.b.uj01dowf1x5lk6lysurbr0lgbdd1wfyw8sm8q17mnt0i9igk774vcwr5rly5dguu/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "illegal character in domain",
-			path:    "a.b!.com/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "invalid all-numeric label",
-			path:    "123/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:  "all-numeric label with param",
-			path:  "123.{a}.456/",
-			wantN: 1,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("123.", true),
-				paramToken("a", ""),
-				staticToken(".456", true),
-				staticToken("/", false),
-			)),
-		},
-		{
-			name:  "all-numeric label with wildcard",
-			path:  "123.+{a}.456/",
-			wantN: 1,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("123.", true),
-				wildcardToken("a", ""),
-				staticToken(".456", true),
-				staticToken("/", false),
-			)),
-		},
-		{
-			name:    "all-numeric label with path wildcard",
-			path:    "123.456/{abc}",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "hostname exceed 255 character",
-			path:    "a.78fayzyiqkt3hh2mquv9szfroeexx8qztscu3oudoyfarjl6jmdyxk2cefvzjx.78fayzyiqkt3hh2mquv9szfroeexx8qztscu3oudoyfarjl6jmdyxk2cefvzjxr.78fayzyiqkt3hh2mquv9szfroeexx8qztscu3oudoyfarjl6jmdyxk2cefvzjxr.78fayzyiqkt3hh2mquv9szfroeexx8qztscu3oudoyfarjl6jmdyxk2cefvzjxr/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "invalid all-numeric label",
-			path:    "11.22.33/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:    "invalid uppercase label",
-			path:    "ABC/",
-			wantErr: ErrInvalidRoute,
-			wantN:   0,
-		},
-		{
-			name:  "2 regular params in domain",
-			path:  "{a}.{b}.com/",
-			wantN: 2,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				paramToken("a", ""),
-				staticToken(".", true),
-				paramToken("b", ""),
-				staticToken(".com", true),
-				staticToken("/", false),
-			)),
-		},
-		{
-			name:  "253 character with .",
-			path:  "78fayzyiqkt3hh2mquv9szfroeexx8qztscu3oudoyfarjl6jmdyxk2cefvzjxr.78fayzyiqkt3hh2mquv9szfroeexx8qztscu3oudoyfarjl6jmdyxk2cefvzjxr.78fayzyiqkt3hh2mquv9szfroeexx8qztscu3oudoyfarjl6jmdyxk2cefvzjxr.78fayzyiqkt3hh2mquv9szfroeexx8qztscu3oudoyfarjl6jmdyxk2cefvzj/",
-			wantN: 0,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("78fayzyiqkt3hh2mquv9szfroeexx8qztscu3oudoyfarjl6jmdyxk2cefvzjxr.78fayzyiqkt3hh2mquv9szfroeexx8qztscu3oudoyfarjl6jmdyxk2cefvzjxr.78fayzyiqkt3hh2mquv9szfroeexx8qztscu3oudoyfarjl6jmdyxk2cefvzjxr.78fayzyiqkt3hh2mquv9szfroeexx8qztscu3oudoyfarjl6jmdyxk2cefvzj", true),
-				staticToken("/", false),
-			)),
-		},
-		{
-			name:  "param does not count at character",
-			path:  "{a}.78fayzyiqkt3hh2mquv9szfroeexx8qztscu3oudoyfarjl6jmdyxk2cefvzjx.78fayzyiqkt3hh2mquv9szfroeexx8qztscu3oudoyfarjl6jmdyxk2cefvzjxr.78fayzyiqkt3hh2mquv9szfroeexx8qztscu3oudoyfarjl6jmdyxk2cefvzjxr.78fayzyiqkt3hh2mquv9szfroeexx8qztscu3oudoyfarjl6jmdyxk2cefvzj/",
-			wantN: 1,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				paramToken("a", ""),
-				staticToken(".78fayzyiqkt3hh2mquv9szfroeexx8qztscu3oudoyfarjl6jmdyxk2cefvzjx.78fayzyiqkt3hh2mquv9szfroeexx8qztscu3oudoyfarjl6jmdyxk2cefvzjxr.78fayzyiqkt3hh2mquv9szfroeexx8qztscu3oudoyfarjl6jmdyxk2cefvzjxr.78fayzyiqkt3hh2mquv9szfroeexx8qztscu3oudoyfarjl6jmdyxk2cefvzj", true),
-				staticToken("/", false),
-			)),
-		},
-		{
-			name:  "hostname variant with multiple catch all suffix and inflight with arg in the middle of the route",
-			path:  "example.com/foo/bar/+{bar}/x+{args}/y/+{z}/{b}",
-			wantN: 4,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("example.com", true),
-				staticToken("/foo/bar/", false),
-				wildcardToken("bar", ""),
-				staticToken("/x", false),
-				wildcardToken("args", ""),
-				staticToken("/y/", false),
-				wildcardToken("z", ""),
-				staticToken("/", false),
-				paramToken("b", ""),
-			)),
-		},
-		{
-			name:  "hostname variant with inflight catch all with arg in the middle of the route",
-			path:  "example.com/foo/bar/damn+{bar}/baz",
-			wantN: 1,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("example.com", true),
-				staticToken("/foo/bar/damn", false),
-				wildcardToken("bar", ""),
-				staticToken("/baz", false),
-			)),
-		},
-		{
-			name:  "hostname variant catch all with arg in the middle of the route and param after",
-			path:  "example.com/foo/bar/+{bar}/{baz}",
-			wantN: 2,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("example.com", true),
-				staticToken("/foo/bar/", false),
-				wildcardToken("bar", ""),
-				staticToken("/", false),
-				paramToken("baz", ""),
-			)),
-		},
-		{
-			name:  "complex domain and path",
-			path:  "{ab}.{c}.de{f}.com/foo/bar/+{bar}/x+{args}/y/+{z}/{b}",
-			wantN: 7,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				paramToken("ab", ""),
-				staticToken(".", true),
-				paramToken("c", ""),
-				staticToken(".de", true),
-				paramToken("f", ""),
-				staticToken(".com", true),
-				staticToken("/foo/bar/", false),
-				wildcardToken("bar", ""),
-				staticToken("/x", false),
-				wildcardToken("args", ""),
-				staticToken("/y/", false),
-				wildcardToken("z", ""),
-				staticToken("/", false),
-				paramToken("b", ""),
-			)),
-		},
-		// Reject path with traversal pattern
-		{
-			name:    "path with double slash",
-			path:    "/foo//bar",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "path with > double slash",
-			path:    "/foo///bar",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "path with slash dot slash",
-			path:    "/foo/./bar",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "path with slash dot slash",
-			path:    "/foo/././bar",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "path with double dot parent reference",
-			path:    "/foo/../bar",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "path with double dot parent reference",
-			path:    "/foo/../../bar",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "path ending with slash dot",
-			path:    "/foo/.",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "path ending with slash double dot",
-			path:    "/foo/..",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "path ending with slash dot",
-			path:    "/.",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "path ending with slash double dot",
-			path:    "/..",
-			wantErr: ErrInvalidRoute,
-		},
-		// Allowed dot and slash combinaison
-		{
-			name: "last path segment starting with slash dot and text",
-			path: "/foo/.bar",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/.bar", false),
-			)),
-		},
-		{
-			name: "last path segment starting with slash dot and text",
-			path: "/foo/..bar",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/..bar", false),
-			)),
-		},
-		{
-			name: "path segment starting with slash dot and text",
-			path: "/foo/.bar/baz",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/.bar/baz", false),
-			)),
-		},
-		{
-			name:  "path segment starting with slash dot and param",
-			path:  "/foo/.{foo}/baz",
-			wantN: 1,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/.", false),
-				paramToken("foo", ""),
-				staticToken("/baz", false),
-			)),
-		},
-		{
-			name: "path segment starting with slash dot and text",
-			path: "/foo/..bar/baz",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/..bar/baz", false),
-			)),
-		},
-		{
-			name:  "path segment starting with slash dot and param",
-			path:  "/foo/..{foo}/baz",
-			wantN: 1,
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/..", false),
-				paramToken("foo", ""),
-				staticToken("/baz", false),
-			)),
-		},
-		{
-			name: "path segment ending with dot slash",
-			path: "/foo/bar./baz",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/bar./baz", false),
-			)),
-		},
-		{
-			name: "path segment ending with double dot slash",
-			path: "/foo/bar../baz",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/bar../baz", false),
-			)),
-		},
-		{
-			name: "path segment with > double dot",
-			path: "/foo/.../baz",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/.../baz", false),
-			)),
-		},
-		{
-			name: "path segment ending with slash and > double dot",
-			path: "/foo/...",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/...", false),
-			)),
-		},
-		{
-			name: "last path segment ending with dot",
-			path: "/foo/bar.",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/bar.", false),
-			)),
-		},
-		{
-			name: "last path segment ending with double dot",
-			path: "/foo/bar..",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/bar..", false),
-			)),
-		},
-		{
-			name: "path segment with dot",
-			path: "/foo/a.b.c",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/a.b.c", false),
-			)),
-		},
-		{
-			name: "path segment with double dot",
-			path: "/foo/a..b..c",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/a..b..c", false),
-			)),
-		},
-		// Regexp
-		{
-			name: "simple ending param with regexp",
-			path: "/foo/{bar:[A-z]+}",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/", false),
-				paramToken("bar", "[A-z]+"),
-			)),
-			wantN: 1,
-		},
-		{
-			name: "simple ending param with regexp",
-			path: "/foo/+{bar:[A-z]+}",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/", false),
-				wildcardToken("bar", "[A-z]+"),
-			)),
-			wantN:             1,
-			wantStartCatchAll: 5,
-		},
-		{
-			name: "simple infix param with regexp",
-			path: "/foo/{bar:[A-z]+}/baz",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/", false),
-				paramToken("bar", "[A-z]+"),
-				staticToken("/baz", false),
-			)),
-			wantN: 1,
-		},
-		{
-			name: "multi infix and ending param with regexp",
-			path: "/foo/{bar:[A-z]+}/{baz:[0-9]+}",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/", false),
-				paramToken("bar", "[A-z]+"),
-				staticToken("/", false),
-				paramToken("baz", "[0-9]+"),
-			)),
-			wantN: 2,
-		},
-		{
-			name: "multi infix and ending wildcard with regexp",
-			path: "/foo/+{bar:[A-z]+}/a+{baz:[0-9]+}",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/", false),
-				wildcardToken("bar", "[A-z]+"),
-				staticToken("/a", false),
-				wildcardToken("baz", "[0-9]+"),
-			)),
-			wantN:             2,
-			wantStartCatchAll: 20,
-		},
-		{
-			name: "consecutive infix regexp wildcard and regexp param allowed",
-			path: "/foo/+{bar:[A-z]+}/{baz:[0-9]+}",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/", false),
-				wildcardToken("bar", "[A-z]+"),
-				staticToken("/", false),
-				paramToken("baz", "[0-9]+"),
-			)),
-			wantN: 2,
-		},
-		{
-			name: "hostname starting with regexp",
-			path: "{a:[A-z]+}.b.c/foo",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				paramToken("a", "[A-z]+"),
-				staticToken(".b.c", true),
-				staticToken("/foo", false),
-			)),
-			wantN: 1,
-		},
-		{
-			name: "hostname with middle param regexp",
-			path: "a.{b:[A-z]+}.c/foo",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("a.", true),
-				paramToken("b", "[A-z]+"),
-				staticToken(".c", true),
-				staticToken("/foo", false),
-			)),
-			wantN: 1,
-		},
-		{
-			name: "hostname ending with param regexp",
-			path: "a.b.{c:[A-z]+}/foo",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("a.b.", true),
-				paramToken("c", "[A-z]+"),
-				staticToken("/foo", false),
-			)),
-			wantN: 1,
-		},
-		{
-			name: "non capturing group allowed in regexp",
-			path: "/foo/{bar:(?:foo|bar)}",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/", false),
-				paramToken("bar", "(?:foo|bar)"),
-			)),
-			wantN: 1,
-		},
-		{
-			name: "regexp wildcard at the beginning of the path",
-			path: "/+{foo:[A-z]+}/bar",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/", false),
-				wildcardToken("foo", "[A-z]+"),
-				staticToken("/bar", false),
-			)),
-			wantN: 1,
-		},
-		{
-			name: "regexp wildcard at the beginning of the host",
-			path: "+{a:[A-z]+}.b.c/",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				wildcardToken("a", "[A-z]+"),
-				staticToken(".b.c", true),
-				staticToken("/", false),
-			)),
-			wantN: 1,
-		},
-		{
-			name: "consecutive wildcard from hostname to path",
-			path: "+{foo}/+{bar}",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				wildcardToken("foo", ""),
-				staticToken("/", false),
-				wildcardToken("bar", ""),
-			)),
-			wantN:             2,
-			wantStartCatchAll: 7,
-		},
-		{
-			name: "consecutive wildcard with empty catch all from hostname to path",
-			path: "+{foo}/*{bar}",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				wildcardToken("foo", ""),
-				staticToken("/", false),
-				wildcardToken("bar", ""),
-			)),
-			wantN:             2,
-			wantStartCatchAll: 7,
-		},
-		{
-			name: "param then wildcard regexp",
-			path: "{a}.+{b:b}/",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				paramToken("a", ""),
-				staticToken(".", true),
-				wildcardToken("b", "b"),
-				staticToken("/", false),
-			)),
-			wantN: 2,
-		},
-		{
-			name: "param regexp then wildcard regexp",
-			path: "{a:a}.+{b:b}/",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				paramToken("a", "a"),
-				staticToken(".", true),
-				wildcardToken("b", "b"),
-				staticToken("/", false),
-			)),
-			wantN: 2,
-		},
-		{
-			name: "catch all empty as suffix",
-			path: "/foo/*{any}",
-			wantTokens: slices.Collect(iterutil.SeqOf(
-				staticToken("/foo/", false),
-				wildcardToken("any", ""),
-			)),
-			wantN:             1,
-			wantStartCatchAll: 5,
-		},
-		{
-			name:    "consecutive infix wildcard at start with regexp not allowed",
-			path:    "/+{foo:[A-z]+}/+{baz:[0-9]+}",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "consecutive wildcard with catch all empty not allowed",
-			path:    "/+{foo}/*{baz}",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "consecutive infix wildcard with catch all empty at start with regexp not allowed",
-			path:    "/+{foo:[A-z]+}/*{baz:[0-9]+}",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "hostname consecutive infix wildcard at start with regexp not allowed",
-			path:    "/{foo:[A-z]+}.+{baz:[0-9]+}/",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "consecutive infix wildcard at start with and without regexp not allowed",
-			path:    "/+{foo:[A-z]+}/+{baz}",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "hostname consecutive infix wildcard at start with and without regexp not allowed",
-			path:    "+{foo:[A-z]+}.+{baz}/",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "consecutive infix wildcard at start with regexp not allowed",
-			path:    "/+{foo}/+{baz:[0-9]+}/",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "hostname consecutive infix wildcard at start with regexp not allowed",
-			path:    "+{foo}.+{baz:[0-9]+}/",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "consecutive infix wildcard with regexp not allowed",
-			path:    "/foo/+{bar:[A-z]+}/+{baz:[0-9]+}",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "hostname consecutive infix wildcard with regexp not allowed",
-			path:    "foo.+{bar:[A-z]+}.+{baz:[0-9]+}/",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "consecutive infix wildcard with first regexp not allowed",
-			path:    "/foo/+{bar:[A-z]+}/+{baz}",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "hostname consecutive infix wildcard with first regexp not allowed",
-			path:    "foo.+{bar:[A-z]+}.+{baz}/",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "consecutive infix wildcard with second regexp not allowed",
-			path:    "/foo/+{bar}/+{baz:[A-z]+}/",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "hostname consecutive infix wildcard with second regexp not allowed",
-			path:    "foo.+{bar}.+{baz:[A-z]+}/",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "non slash char after regexp param not allowed",
-			path:    "/foo/{bar:[A-z]+}a/",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "non slash char after regexp wildcard not allowed",
-			path:    "/foo/+{bar:[A-z]+}a/",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "regexp wildcard not allowed in hostname",
-			path:    "+{a.{b:[A-z]+}}.c/",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "regexp wildcard not allowed in hostname",
-			path:    "+{a.b.{c:[A-z]+}/",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "missing param name with regexp",
-			path:    "/foo/{:[A-z]+}",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "missing wildcard name with regexp",
-			path:    "/foo/+{:[A-z]+}",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "missing regular expression",
-			path:    "/foo/{a:}",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "missing regular expression with only ':'",
-			path:    "/foo/{:}",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "unsupported regexp in optional wildcard",
-			path:    "/foo/*{any:[A-z]+}",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "unbalanced braces in param regexp",
-			path:    "/foo/{bar:[A-z]+",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "unbalanced braces in wildcard regexp",
-			path:    "/foo/+{bar:[A-z]+",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "balanced braces in param regexp with invalid char after",
-			path:    "/foo/{bar:{}}a",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "balanced braces in wildcard regexp with invalid brace after",
-			path:    "/foo/{bar:{}}}",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "unbalanced braces in regexp complex",
-			path:    "/foo/{bar:{{{{}}}}",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "invalid regular expression",
-			path:    "/foo/{bar:a{5,2}}",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "invalid regular expression",
-			path:    "/foo/{bar:\\k}",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "capture group in regexp are not allowed",
-			path:    "/foo/{bar:(foo|bar)}",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "no opening brace after * wildcard",
-			path:    "/foo/*:bar}",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "no infix catch all empty",
-			path:    "/foo/*{any}/bar",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "no infix inflight catch all empty",
-			path:    "/foo/uuid_*{any}/bar",
-			wantErr: ErrInvalidRoute,
-		},
-		{
-			name:    "no suffix catch all empty in hostname",
-			path:    "a.b.*{any}/",
-			wantErr: ErrInvalidRoute,
-		},
-	}
-	for _, tc := range cases {
-		t.Run(tc.name, func(t *testing.T) {
-			parsed, err := f.parseRoute(tc.path)
-			require.ErrorIs(t, err, tc.wantErr)
-			assert.Equal(t, tc.wantN, parsed.paramCnt)
-			assert.Equal(t, tc.wantTokens, parsed.token)
-			assert.Equal(t, tc.wantStartCatchAll, parsed.startCatchAll)
-			if err == nil {
-				assert.Equal(t, strings.IndexByte(tc.path, '/'), parsed.endHost)
-			}
-		})
-	}
-}
-
-func TestParseRouteParamsConstraint(t *testing.T) {
-	t.Run("param limit", func(t *testing.T) {
-		f, _ := NewRouter(WithMaxRouteParams(3))
-		_, err := f.parseRoute("/{1}/{2}/{3}")
-		assert.NoError(t, err)
-		_, err = f.parseRoute("/{1}/{2}/{3}/{4}")
-		assert.Error(t, err)
-		_, err = f.parseRoute("/ab{1}/{2}/cd/{3}/{4}/ef")
-		assert.Error(t, err)
-	})
-	t.Run("param key limit", func(t *testing.T) {
-		f, _ := NewRouter(WithMaxRouteParamKeyBytes(3))
-		_, err := f.parseRoute("/{abc}/{abc}/{abc}")
-		assert.NoError(t, err)
-		_, err = f.parseRoute("/{abcd}/{abc}/{abc}")
-		assert.Error(t, err)
-		_, err = f.parseRoute("/{abc}/{abcd}/{abc}")
-		assert.Error(t, err)
-		_, err = f.parseRoute("/{abc}/{abc}/{abcd}")
-		assert.Error(t, err)
-		_, err = f.parseRoute("/{abc}/+{abcd}/{abc}")
-		assert.Error(t, err)
-		_, err = f.parseRoute("/{abc}/{abc}/+{abcdef}")
-		assert.Error(t, err)
-	})
-	t.Run("param key limit with regexp", func(t *testing.T) {
-		f, _ := NewRouter(WithMaxRouteParamKeyBytes(3), AllowRegexpParam(true))
-		_, err := f.parseRoute("/{abc:a}/{abc:a}/{abc:a}")
-		assert.NoError(t, err)
-		_, err = f.parseRoute("/{abcd:a}/{abc:a}/{abc:a}")
-		assert.Error(t, err)
-		_, err = f.parseRoute("/{abc:a}/{abcd:a}/{abc:a}")
-		assert.Error(t, err)
-		_, err = f.parseRoute("/{abc:a}/{abc:a}/{abcd:a}")
-		assert.Error(t, err)
-		_, err = f.parseRoute("/{abc:a}/+{abcd:a}/{abc:a}")
-		assert.Error(t, err)
-		_, err = f.parseRoute("/{abc:a}/{abc:a}/+{abcdef:a}")
-		assert.Error(t, err)
-	})
-	t.Run("disabled regexp support for param", func(t *testing.T) {
-		f, _ := NewRouter()
-		_, err := f.parseRoute("/{a}/{b}/{c}")
-		assert.NoError(t, err)
-		// path params
-		_, err = f.parseRoute("/{a:a}/{b}/{c}")
-		assert.Error(t, err)
-		_, err = f.parseRoute("/{a}/{b:b}/{c}")
-		assert.Error(t, err)
-		_, err = f.parseRoute("/{a}/{b}/{c:c}")
-		assert.Error(t, err)
-		// hostname params
-		_, err = f.parseRoute("{a:a}.{b}.{c}/")
-		assert.Error(t, err)
-		_, err = f.parseRoute("{a}.{b:b}.{c}/")
-		assert.Error(t, err)
-		_, err = f.parseRoute("{a}.{b}.{c:c}/")
-		assert.Error(t, err)
-	})
-	t.Run("disabled regexp support for wildcard", func(t *testing.T) {
-		f, _ := NewRouter()
-		_, err := f.parseRoute("/{a}/{b}/{c}")
-		assert.NoError(t, err)
-		// wildcard
-		_, err = f.parseRoute("/+{a:a}/{b}/{c}")
-		assert.Error(t, err)
-		_, err = f.parseRoute("/{a}/+{b:b}/{c}")
-		assert.Error(t, err)
-		_, err = f.parseRoute("/{a}/{b}/+{c:c}")
-		assert.Error(t, err)
-	})
-}
-
 func TestRouteMatchersConstraint(t *testing.T) {
 	t.Run("insert: enforce max route matchers", func(t *testing.T) {
 		f, _ := NewRouter(WithMaxRouteMatchers(3))
